Replace uint8 with uint256

+++ contracts/applications/NewPropertyManager.sol
@@ -307,7 +307,7 @@ contract NewPropertyManager is AbstractOracleApplication {
   }
 
   function assignLockedStatus(bytes32 _aId) internal {
-    for (uint8 i = 0; i < applications[_aId].assignedOracleTypes.length; i++) {
+    for (uint256 i = 0; i < applications[_aId].assignedOracleTypes.length; i++) {
       if (applications[_aId].validationStatus[applications[_aId].assignedOracleTypes[i]] != ValidationStatus.LOCKED) {
         changeValidationStatus(applications[_aId], applications[_aId].assignedOracleTypes[i], ValidationStatus.LOCKED);
       }
@@ -394,7 +394,7 @@ contract NewPropertyManager is AbstractOracleApplication {
     uint256 len = a.assignedOracleTypes.length;
     bool allApproved = true;
 
-    for (uint8 i = 0; i < len; i++) {
+    for (uint256 i = 0; i < len; i++) {
       if (a.validationStatus[a.assignedOracleTypes[i]] != ValidationStatus.APPROVED) {
         allApproved = false;
       }
@@ -466,7 +466,7 @@ contract NewPropertyManager is AbstractOracleApplication {
     requireOracleActiveWithAssignedActiveOracleType(a.pgg, msg.sender, senderOracleType);
     require(a.validationStatus[senderOracleType] == ValidationStatus.LOCKED, "Application should be locked first");
 
-    for (uint8 i = 0; i < len; i++) {
+    for (uint256 i = 0; i < len; i++) {
       if (a.validationStatus[a.assignedOracleTypes[i]] == ValidationStatus.PENDING) {
         revert("All oracle types should lock the application first");
       }
